Further enforce the use of VhdlName
