+++ crates/vhdl/src/object/record.rs
@@ -3,11 +3,14 @@ use std::iter::FromIterator;
 use indexmap::IndexMap;
 use textwrap::indent;
 use tydi_common::error::{Error, Result};
+use tydi_common::traits::Identify;
 
 use crate::architecture::arch_storage::Arch;
-use crate::common::vhdl_name::VhdlName;
+use crate::common::vhdl_name::{VhdlName, VhdlNameSelf};
 use crate::declaration::DeclareWithIndent;
-use crate::object::ObjectType;
+use crate::object::object_type::ObjectType;
+
+use super::object_type::DeclarationTypeName;
 
 /// A record object
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
@@ -50,10 +53,6 @@ impl RecordObject {
         }
     }
 
-    pub fn type_name(&self) -> String {
-        self.type_name.to_string()
-    }
-
     pub fn fields(&self) -> IndexMap<VhdlName, &ObjectType> {
         IndexMap::from_iter(self.fields.iter().map(|x| (x.field().clone(), x.typ())))
     }
@@ -64,7 +63,7 @@ impl RecordObject {
             None => Err(Error::InvalidArgument(format!(
                 "Field {} does not exist on record with type {}",
                 field_name.to_string(),
-                self.type_name()
+                self.vhdl_name()
             ))),
         }
     }
@@ -72,13 +71,31 @@ impl RecordObject {
 
 impl DeclareWithIndent for RecordObject {
     fn declare_with_indent(&self, _db: &dyn Arch, indent_style: &str) -> Result<String> {
-        let mut this = format!("type {} is record\n", self.type_name());
+        let mut this = format!("type {} is record\n", self.vhdl_name());
         let mut fields = String::new();
         for (name, typ) in self.fields() {
-            fields.push_str(format!("{} : {};\n", name, typ.type_name()).as_str());
+            fields.push_str(format!("{} : {};\n", name, typ.declaration_type_name()).as_str());
         }
         this.push_str(&indent(&fields, indent_style));
         this.push_str("end record;");
         Ok(this)
+    }
+}
+
+impl VhdlNameSelf for RecordObject {
+    fn vhdl_name(&self) -> &VhdlName {
+        &self.type_name
+    }
+}
+
+impl Identify for RecordObject {
+    fn identifier(&self) -> String {
+        self.vhdl_name().to_string()
+    }
+}
+
+impl DeclarationTypeName for RecordObject {
+    fn declaration_type_name(&self) -> String {
+        self.identifier()
     }
 }
